refactor(ui): split sandboxClient.ts god object into modules
