+++ scripts/.god-object-baseline.txt
@@ -14,5 +14,4 @@
 682 ui/src/components/shared/InfrastructureModal.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 627 ui/src/lib/hooks/useCreateDeploy.ts
-624 ui/src/lib/api/sandboxClient.ts
 606 sandbox-runtime/src/util.rs
